Moved base - axis tests into separate file
